+++ lib/compiler-singlepass/src/machine_x64.rs
@@ -676,21 +676,12 @@ impl MachineX86_64 {
         );
 
         self.emit_label(trap_overflow);
-        let offset = self.assembler.get_offset().0;
-        self.trap_table
-            .offset_to_code
-            .insert(offset, TrapCode::IntegerOverflow);
-        self.emit_illegal_op();
-        self.mark_instruction_address_end(offset);
+
+        self.emit_illegal_op_internal(TrapCode::IntegerOverflow);
 
         self.emit_label(trap_badconv);
 
-        let offset = self.assembler.get_offset().0;
-        self.trap_table
-            .offset_to_code
-            .insert(offset, TrapCode::BadConversionToInteger);
-        self.emit_illegal_op();
-        self.mark_instruction_address_end(offset);
+        self.emit_illegal_op_internal(TrapCode::BadConversionToInteger);
 
         self.emit_label(end);
     }
@@ -819,20 +810,10 @@ impl MachineX86_64 {
         );
 
         self.emit_label(trap_overflow);
-        let offset = self.assembler.get_offset().0;
-        self.trap_table
-            .offset_to_code
-            .insert(offset, TrapCode::IntegerOverflow);
-        self.emit_illegal_op();
-        self.mark_instruction_address_end(offset);
+        self.emit_illegal_op_internal(TrapCode::IntegerOverflow);
 
         self.emit_label(trap_badconv);
-        let offset = self.assembler.get_offset().0;
-        self.trap_table
-            .offset_to_code
-            .insert(offset, TrapCode::BadConversionToInteger);
-        self.emit_illegal_op();
-        self.mark_instruction_address_end(offset);
+        self.emit_illegal_op_internal(TrapCode::BadConversionToInteger);
 
         self.emit_label(end);
     }
@@ -1670,6 +1651,10 @@ impl MachineX86_64 {
     fn emit_unwind_op(&mut self, op: UnwindOps) {
         self.unwind_ops.push((self.get_offset().0, op));
     }
+    fn emit_illegal_op_internal(&mut self, trap: TrapCode) {
+        let v = trap as u8;
+        self.assembler.emit_ud1_payload(v);
+    }
 }
 
 impl Machine for MachineX86_64 {
@@ -2331,8 +2316,22 @@ impl Machine for MachineX86_64 {
         self.release_simd(tmp1);
     }
 
-    fn emit_illegal_op(&mut self) {
+    fn emit_illegal_op(&mut self, trap: TrapCode) {
+        // code below is kept as a reference on how to emit illegal op with trap info
+        // without an Undefined opcode with payload
+        /*
+        let offset = self.assembler.get_offset().0;
+        self.trap_table
+        .offset_to_code
+        .insert(offset, trap);
         self.assembler.emit_ud2();
+        self.mark_instruction_address_end(offset);*/
+        let v = trap as u8;
+        // payload needs to be between 0-15
+        // this will emit an 40 0F B9 Cx opcode, with x the payload
+        let offset = self.assembler.get_offset().0;
+        self.assembler.emit_ud1_payload(v);
+        self.mark_instruction_address_end(offset);
     }
     fn get_label(&mut self) -> Label {
         self.assembler.new_dynamic_label()