+++ lib/wasi/src/syscalls/unix/mod.rs
@@ -8,7 +8,6 @@ use wasmer::WasmRef;
 
 pub fn platform_clock_res_get(
     clock_id: __wasi_clockid_t,
-    resolution: WasmRef<__wasi_timestamp_t>,
 ) -> Result<i64, __wasi_errno_t> {
     let unix_clock_id = match clock_id {
         __WASI_CLOCK_MONOTONIC => CLOCK_MONOTONIC,